@@ -19,21 +19,24 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@monaco-editor/loader": "^1.3.2",
-    "antd": "^4.x",
+    "ahooks": "^3.7.4",
+    "antd": "^5.0.5",
     "axios": "^0.21.1",
+    "dayjs": "^1.11.7",
+    "echarts": "^5.4.0",
+    "moment": "^2.29.4",
     "monaco-editor": "^0.33.0",
     "npm": "^8.19.2",
-    "react": "^16.8.0",
+    "ramda": "^0.28.0",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-intl": "^5.25.1",
+    "react-keep-alive": "^2.5.2",
     "react-router-dom": "^5.2.0"
   },
-  "peerDependencies": {
-    "antd": "^4.x",
-    "react": "^16.8.0"
-  },
   "devDependencies": {
+    "@ant-design/moment-webpack-plugin": "^0.0.2",
     "@babel/core": "^7.17.10",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-runtime": "^7.17.10",
@@ -42,8 +45,9 @@
     "@babel/preset-typescript": "^7.16.7",
     "@babel/runtime-corejs3": "^7.17.9",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.5",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/ramda": "^0.28.22",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/react-redux": "^7.1.24",
     "@types/react-router-dom": "^5.2.0",
     "@types/redux-logger": "^3.0.9",
