@@ -13,6 +13,7 @@ import useRequiredRule from '../../../hooks/useRequiredRule';
 import { TDetail, TEngineParameter } from '../type';
 import './index.less';
 import store, { RootReducer } from '@/store';
+
 type InnerProps = {
     form: FormInstance,
     detail: TDetail
@@ -24,9 +25,8 @@ const Index = ({ form, detail }: InnerProps) => {
     const requiredRule = useRequiredRule();
     const [engineList, setEngineList] = useState([]);
     const { datasourceReducer } = store.getState() as RootReducer;
-    console.log('123123',detail)
+    // console.log('123123', detail);
     useMount(async () => {
-       
         try {
             const $engineList = await $http.get('metric/engine/list');
             setEngineList($engineList || []);
@@ -41,10 +41,10 @@ const Index = ({ form, detail }: InnerProps) => {
                 others: paramter.others ?? '--conf spark.yarn.maxAppAttempts=1',
                 tenantCode: detail?.tenantCode,
                 env: detail?.env,
-                engineType: detail?.engineType || (datasourceReducer.modeType === 'comparison' ? 'spark' : 'jdbc')
+                engineType: detail?.engineType || (datasourceReducer.modeType === 'comparison' ? 'spark' : 'local'),
             });
         } catch (error) {
-            console.log('error',error)
+            console.log('error', error);
         }
     });
     const renderSpark = () => (
@@ -139,9 +139,10 @@ const Index = ({ form, detail }: InnerProps) => {
                             source={engineList}
                             sourceValueMap="key"
                             style={{ width: 200 }}
-                            disabled={ datasourceReducer.modeType === 'comparison' }
+                            disabled={datasourceReducer.modeType === 'comparison'}
                         />
                     </Form.Item>
+
                 </Col>
             </Row>
             <Form.Item noStyle dependencies={['engineType']}>
