@@ -1,9 +1,11 @@
 import React, { useState, useImperativeHandle } from 'react';
 import {
-    Row, Col, Form, Input, FormInstance, Button
+    Row, Col, Form, Input, FormInstance, Button,
 } from 'antd';
 import './index.less';
 import { useIntl } from 'react-intl';
+import TextArea from 'antd/lib/input/TextArea';
+import { MinusCircleOutlined } from '@ant-design/icons';
 import { layoutItem, layoutOneLineItem, layoutNoneItem } from '../helper';
 import useRequest from '../../../hooks/useRequest';
 import useRequiredRule from '../../../hooks/useRequiredRule';
@@ -13,9 +15,7 @@ import {
     CustomSelect, useMount, usePersistFn, IF,
 } from '../../../common';
 import Title from '../Title';
-import TextArea from 'antd/lib/input/TextArea';
 import store, { RootReducer } from '@/store';
-import { MinusCircleOutlined } from '@ant-design/icons';
 
 type InnerProps = {
     form: FormInstance,
@@ -30,7 +30,7 @@ type dynamicConfigItem = {
 }
 
 const Index = ({
-    form, id, detail, setMetricTypeParent
+    form, id, detail, setMetricTypeParent,
 }: InnerProps) => {
     const intl = useIntl();
     const { $http } = useRequest();
@@ -45,11 +45,11 @@ const Index = ({
     const [column2, setCloumn2] = useState([]);
     useMount(async () => {
         try {
-            getDatabases(id, 1, true)
+            getDatabases(id, 1, true);
             if (!detail || !detail.id) {
                 form.setFieldsValue({
                     mappingColumns: [{}],
-                    dataSourceId: parseInt(id)
+                    dataSourceId: parseInt(id),
                 });
             } else {
                 form.setFieldsValue({
@@ -58,16 +58,16 @@ const Index = ({
                     metricType: detail?.metricType,
                     metricParameter: detail?.parameterItem?.metricParameter,
                     metricParameter2: detail?.parameterItem?.metricParameter2,
-                    mappingColumns: detail?.parameterItem?.mappingColumns
+                    mappingColumns: detail?.parameterItem?.mappingColumns,
                 });
                 // console.log('detail', detail)
-                setMetricTypeParent(detail?.metricType)
-                setMetricType(detail?.metricType)
-                getDatabases(detail?.dataSourceId2, 2, true)
-                getTable(detail?.parameterItem?.metricParameter?.database, detail?.dataSourceId2, 1, true)
-                getTable(detail?.parameterItem?.metricParameter2?.database2, detail?.dataSourceId2, 2, true)
-                getCloumn(detail?.parameterItem?.metricParameter?.table, id,detail?.parameterItem?.metricParameter?.database, 1, false)
-                getCloumn(detail?.parameterItem?.metricParameter2?.table2, detail?.dataSourceId2, detail?.parameterItem?.metricParameter2?.database2, 2, false)
+                setMetricTypeParent(detail?.metricType);
+                setMetricType(detail?.metricType);
+                getDatabases(detail?.dataSourceId2, 2, true);
+                getTable(detail?.parameterItem?.metricParameter?.database, detail?.dataSourceId2, 1, true);
+                getTable(detail?.parameterItem?.metricParameter2?.database2, detail?.dataSourceId2, 2, true);
+                getCloumn(detail?.parameterItem?.metricParameter?.table, id, detail?.parameterItem?.metricParameter?.database, 1, false);
+                getCloumn(detail?.parameterItem?.metricParameter2?.table2, detail?.dataSourceId2, detail?.parameterItem?.metricParameter2?.database2, 2, false);
             }
 
             const $metricList = await $http.get('metric/list/DATA_RECONCILIATION');
@@ -78,50 +78,51 @@ const Index = ({
                 workSpaceId: workspaceId,
                 pageNumber: 1,
                 pageSize: 9999,
-            })
+            });
+            // console.log('执行');
             setDataSoucre($dataSoucrce.records || []);
         } catch (error) {
-            console.log('erro', error)
+            console.log('erro', error);
         }
     });
     const changeRule = (val: string) => {
-        setMetricTypeParent(val)
-        setMetricType(val)
-    }
+        setMetricTypeParent(val);
+        setMetricType(val);
+    };
     const getDatabases = async (id: string | undefined, index: number, isInit: boolean | undefined) => {
-        if (!id) return
-        const $databases = await $http.get(`datasource/${id}/databases`)
+        if (!id) return;
+        const $databases = await $http.get(`datasource/${id}/databases`);
         if (index === 1) {
-            setDatabases1($databases)
-            if (isInit) return
-            setTable1([])
-            setCloumn1([])
+            setDatabases1($databases);
+            if (isInit) return;
+            setTable1([]);
+            setCloumn1([]);
         } else {
-            setDatabases2($databases)
-            if (isInit) return
-            setTable2([])
-            setCloumn2([])
+            setDatabases2($databases);
+            if (isInit) return;
+            setTable2([]);
+            setCloumn2([]);
         }
-    }
+    };
     const getTable = async (database: string | undefined, id: string | undefined, index: number, isInit: boolean | undefined) => {
-        if (!id || !database) return
-        const $table = await $http.get(`datasource/${id}/${database}/tables`)
+        if (!id || !database) return;
+        const $table = await $http.get(`datasource/${id}/${database}/tables`);
         if (index === 1) {
-            setTable1($table)
-            isInit && setCloumn1([])
+            setTable1($table);
+            isInit && setCloumn1([]);
         } else {
-            setTable2($table)
-            isInit && setCloumn2([])
+            setTable2($table);
+            isInit && setCloumn2([]);
         }
-    }
+    };
     const getCloumn = async (table: string | undefined, id: string | undefined, database: string | undefined, index: number, isInit: boolean | undefined) => {
-        if (!table || !id || !database) return
-        const $column = await $http.get(`datasource/${id}/${database}/${table}/columns`)
-        index === 1 ? setCloumn1($column.columns || []) : setCloumn2($column.columns || [])
-    }
+        if (!table || !id || !database) return;
+        const $column = await $http.get(`datasource/${id}/${database}/${table}/columns`);
+        index === 1 ? setCloumn1($column.columns || []) : setCloumn2($column.columns || []);
+    };
     return (
         <div style={{ padding: '0 10px' }}>
-            <Row gutter={30} >
+            <Row gutter={30}>
                 <Col span={12}>
                     <Form.Item
                         {...layoutItem}
@@ -250,13 +251,19 @@ const Index = ({
                         </Form.Item>
                     </Col>
                 </Row>
-                <Form.List name="mappingColumns" >
+                <Form.List name="mappingColumns">
                     {(fields, { add, remove }, { errors }) => (
                         <>
                             <Row gutter={30}>
                                 <Col span={21} push={3} className="dv-editor-title_flex">
-                                    <span>{intl.formatMessage({ id: 'dv_metric_check_column' })} </span>
-                                    <span style={{ cursor: 'pointer' }} onClick={() => add()}>{intl.formatMessage({ id: 'dv_metric_add' })} </span>
+                                    <span>
+                                        {intl.formatMessage({ id: 'dv_metric_check_column' })}
+                                        {' '}
+                                    </span>
+                                    <span style={{ cursor: 'pointer' }} onClick={() => add()}>
+                                        {intl.formatMessage({ id: 'dv_metric_add' })}
+                                        {' '}
+                                    </span>
                                 </Col>
                             </Row>
                             {
@@ -266,7 +273,7 @@ const Index = ({
                                         key={field.key}
                                         style={{ marginBottom: 0 }}
                                     >
-                                        <Row gutter={30} >
+                                        <Row gutter={30}>
                                             <Col span={9} push={3}>
                                                 <Form.Item
                                                     {...layoutNoneItem}
@@ -311,7 +318,8 @@ const Index = ({
                                             </Col> */}
                                         </Row>
                                     </Form.Item>
-                                ))}
+                                ))
+                            }
                         </>
                     )}
 
