@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Table, Button, Input } from 'antd';
-import { ColumnsType } from 'antd/es/table';
+import { ColumnsType } from 'antd/lib/table';
 import { useIntl } from 'react-intl';
 import { EyeOutlined } from '@ant-design/icons';
-import { ModalProps } from 'antd/es/modal';
+import { ModalProps } from 'antd/lib/modal';
 import {
     Popconfirm, useMount, useModal, useImmutable, useFormRender, IFormRender, useContextModal, IF,
 } from '../../../common';
@@ -118,7 +118,7 @@ const Index: React.FC<IndexProps> = () => {
             render: (text: string, record: IDataSourceListItem) => (
                 <>
                     <a onClick={() => { showFormSchemaModal(record); }}>{intl.formatMessage({ id: 'dv_metric_edit' })}</a>
-                    <IF visible={data?.currentValue != record.id}>
+                    <IF visible={data?.currentValue !== record.id}>
                         <Popconfirm
                             onClick={async () => {
                                 try {
