@@ -5,9 +5,14 @@ import './index.less';
 import enUS from './locale/en_US';
 import zhCN from './locale/zh_CN';
 import { IDvEditorProps } from './type';
-import { EditorProvider, useEditorActions, setEditorFn } from './store/editor';
+import {
+    EditorProvider, useEditorActions, setEditorFn,
+} from './store/editor';
 import { useMount, useWatch } from './common';
 import { MetricConfig } from './components/MetricModal';
+// eslint-disable-next-line import/order
+import querystring from 'querystring';
+import { $http } from '@/http';
 
 const App = (props: IDvEditorProps) => {
     const {
@@ -15,7 +20,31 @@ const App = (props: IDvEditorProps) => {
     } = props;
     const [loading, setLoading] = useState(true);
     const fns = useEditorActions({ setEditorFn });
-    useMount(() => {
+    const [qs] = useState(querystring.parse(window.location.href.split('?')[1] || ''));
+
+    // console.log("qs", qs)
+    // console.log('selectDatabases', JSON.stringify(selectDatabases));
+    useMount(async () => {
+        let { name } = qs;
+        let uuid = '';
+        // if (!name && id) {
+        const $dataSoucrce = await $http.get('/datasource/page', {
+            workSpaceId: workspaceId,
+            pageNumber: 1,
+            pageSize: 9999,
+        });
+        $dataSoucrce.records.forEach((element: {
+                uuid: string; id: string | number; name: string | string[] | undefined;
+}) => {
+            // console.log(element.id === id, element.id, id);
+            if (`${element.id}` === id) {
+                name = element.name;
+                uuid = element.uuid;
+            }
+        });
+        // console.log('执行');
+        // }
+        console.log('uuid');
         fns.setEditorFn({
             workspaceId,
             monacoConfig,
@@ -23,12 +52,39 @@ const App = (props: IDvEditorProps) => {
             baseURL,
             headers: headers || {},
             id,
+            selectDatabases: [{
+                name,
+                uuid,
+            }],
         });
         setLoading(false);
     });
-    useWatch(id, () => {
+
+    useWatch(id, async () => {
+        let { name } = qs;
+        let uuid = '';
+        // if (!name && id) {
+        const $dataSoucrce = await $http.get('/datasource/page', {
+            workSpaceId: workspaceId,
+            pageNumber: 1,
+            pageSize: 9999,
+        });
+        $dataSoucrce.records.forEach((element: {
+                uuid: string; id: string | number; name: string | string[] | undefined;
+}) => {
+            if (element.id === id) {
+                name = element.name;
+                uuid = element.uuid;
+            }
+        });
+        console.log('执行');
+        // }
         fns.setEditorFn({
             id,
+            selectDatabases: [{
+                name,
+                uuid,
+            }],
         });
     });
     if (loading) {
