@@ -1,4 +1,4 @@
 import React, { memo } from 'react';
 import logoPng from 'assets/images/logo-light.png';
 
-export default memo(() => <img className="logo" src={logoPng} />);
+export default memo(({ style }:any) => <img style={style} className="logo" src={logoPng} />);
