@@ -27,15 +27,18 @@ export default {
     common_fail: '操作失败',
     common_updater: '更新人',
     common_back: '返回',
+    common_Logout: '退出系统',
+    common_error_tips: '用户未登录',
 
     '/main/home': '数据源管理',
     '/main/warning': '告警管理',
     '/main/userManage': '用户管理',
     '/main/errorDataManage': '错误数据管理',
-    '/main/detail/:id/editor': '数据开发',
+    '/main/detail/:id/editor': '数据目录',
     '/main/detail/:id/jobs': '作业管理',
     '/main/detail/:id/tasks': '结果管理',
     '/main/detail/:id': '详情',
+    '/main/label': '标签管理',
 
     confirm_text: '确认',
     test_link: '测试链接',
@@ -70,9 +73,12 @@ export default {
 
     header_top_search_msg: '请选择',
     header_top_add_workspace: '新增工作空间',
+    header_top_workspace: '工作空间',
 
     home_search_placeholder: '搜索数据源',
     home_create_datasource: '创建数据源',
+    home_Metadata_scraping: '元数据抓取',
+    home_Metadata_crawl_schedule: '元数据抓取调度',
 
     // jobs
     jobs_list: '作业列表',
@@ -111,6 +117,8 @@ export default {
     jobs_save_run: '保存并运行',
     jobs_tabs_schedule: '定时任务配置',
     jobs_tabs_SLA: 'SLA配置',
+    jobs_add_metirc: '添加作业',
+    jobs_run_Profile: '运行配置文件',
 
     jobs_schedule_type: '调度类型',
     jobs_schedule_cycle: '调度周期',
@@ -151,6 +159,51 @@ export default {
     jobs_Friday: '星期五',
     jobs_Saturday: '星期六',
     jobs_Sunday: '星期日',
+    jobs_Editor: 'SQL编辑器',
+    jobs_Directory: '目录',
+    jobs_Last_Scan_Time: '上次扫描时间',
+    jobs_Number_Of_Labels: '标签数量',
+    jobs_Number_Of_Rules: '规则数量',
+    jobs_Use_heat: '使用热度',
+    job_quantity: '数量',
+    job_Table_name: '表名',
+    job_Column: '列',
+    job_Column_Name: '列名',
+    job_alter: '变更',
+    job_Table: '表',
+    job_database_Name: '数据库名',
+    job_database: '数据库',
+    job_Last_Refrsh_time: '最后刷新时间',
+    job_Metrics: '作业',
+    job_Create_Time: '创建时间',
+    job_Title: '标题',
+    job_Last_Update_Time: '最后更新时间',
+    job_null: '空值',
+    job_notNull: '不是空值',
+    job_Unique: '独有',
+    job_Distinct: '不同',
+    job_Comment: '评论',
+    job_Status: '状态',
+    job_Trend: '趋势',
+    job_Action: '运行',
+    job_Before: '之前',
+    job_After: '之后',
+    job_Schema_Changes: '架构更改',
+    job_Profile: '概要',
+    job_issue: '问题',
+    job_issue_detail: '问题详情',
+    job_good: '好的',
+    job_bad: '坏的',
+    job_Maximum: '最大',
+    job_Minmum: '最小',
+    job_Max_Length: '最大长度',
+    job_Min_Length: '最短长度',
+    job_Avg_Length: '平均长度',
+    job_Blank: '空白',
+    job_Sum: '合计',
+    job_Avg: '平均',
+    job_StdDev: '标准开发',
+    job_Varaiance: '方差',
 
     warn_SLAs_name: '名称',
     warn_SLAs_type: '类型',
@@ -197,4 +250,12 @@ export default {
     error_create_btn: '创建错误数据存储',
     error_table_store_name: '存储名称',
     error_table_store_type: '存储类型',
+    error_title: '存储管理',
+
+    user_title: '用户管理',
+    label_title: '标签分类',
+    label_list: '标签列表',
+    label_add_Classification: '新增标签分类',
+    label_add: '新增标签',
+    label_name: '标签名',
 };
