@@ -1,10 +1,13 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
-    Pagination, Card, Dropdown, Menu, Popconfirm, Row, Col,
+    Pagination, Dropdown, Menu, Popconfirm, Row, Col, Tag, Modal,
 } from 'antd';
-import { EyeOutlined, DeleteOutlined } from '@ant-design/icons';
+import {
+    EditOutlined, DeleteOutlined, EllipsisOutlined, FundOutlined,
+} from '@ant-design/icons';
 import { useIntl } from 'react-intl';
 import { IDataSourceListItem, IDataSourceList } from '@/type/dataSource';
+import Schedule from '@/view/Main/HomeDetail/Jobs/components/Schedule';
 
 type IndexProps = {
     tableData: IDataSourceList,
@@ -25,14 +28,123 @@ const Index: React.FC<IndexProps> = ({
             pageSize,
         });
     };
+    const [isScheduleOpen, setisScheduleOpen] = useState(false);
+    const [uuid, setUuid] = useState<string | number>('');
     const renderItem = (item: IDataSourceListItem) => (
         <React.Fragment key={item.id}>
             <Col span={6}>
-                <Card
+                <div style={{
+                    background: '#FFF4D8',
+                    borderRadius: '8px',
+                    padding: '20px',
+                    marginBottom: '20px',
+                }}
+                >
+                    <p style={{
+                        margin: '0px 0px 20px 0px',
+                        color: '#111',
+                        fontWeight: '600',
+                    }}
+                    >
+                        <span
+                            style={{
+                                cursor: 'pointer',
+                            }}
+                            onClick={() => goDetail(item)}
+                        >
+                            {item.name}
+                        </span>
+
+                        <Tag
+                            color="#32D74B"
+                            style={{
+                                marginLeft: '10px',
+                            }}
+                        >
+                            {/* {intl.formatMessage({ id: 'datasource_modal_source_type' })} */}
+                            {item.type}
+                        </Tag>
+                        <Dropdown
+                            overlay={(
+                                <Menu
+                                    items={[
+                                        {
+                                            label: <div style={{ width: '100%', height: '100%' }} onClick={() => (onEdit(item))}>
+                                                <EditOutlined style={{ marginRight: '4px' }} />
+                                                {intl.formatMessage({ id: 'common_edit' })}
+                                            </div>,
+                                            key: 'edit',
+                                            // icon: ,
+                                        },
+                                        {
+                                            label: (
+                                                <Popconfirm
+                                                    title={intl.formatMessage({ id: 'common_delete_tip' })}
+                                                    onConfirm={() => { onDelete(item); }}
+                                                    okText={intl.formatMessage({ id: 'common_Ok' })}
+                                                    cancelText={intl.formatMessage({ id: 'common_Cancel' })}
+                                                >
+                                                    <a style={{ color: '#f81d22' }}>
+                                                        <DeleteOutlined style={{ color: '#f81d22', marginRight: '4px' }} />
+                                                        {intl.formatMessage({ id: 'common_delete' })}
+                                                    </a>
+                                                </Popconfirm>
+                                            ),
+                                            // icon: <DeleteOutlined style={{ color: '#f81d22' }} />,
+                                            key: 'delete',
+                                            // style:{{color: '#f81d22' }
+                                        },
+                                        {
+                                            label: (
+
+                                                <a onClick={() => {
+                                                    setisScheduleOpen(true);
+                                                    setUuid(item.id);
+                                                }}
+                                                >
+                                                    <FundOutlined style={{ marginRight: '4px' }} />
+                                                    {intl.formatMessage({ id: 'home_Metadata_scraping' })}
+                                                </a>
+
+                                            ),
+                                            // icon: <DeleteOutlined style={{ color: '#f81d22' }} />,
+                                            key: 'push',
+                                            // style:{{color: '#f81d22' }
+                                        },
+                                    ]}
+                                />
+                            )}
+                            placement="bottomLeft"
+                        >
+                            <EllipsisOutlined style={{
+                                float: 'right',
+                                color: '#000',
+                            }}
+                            />
+                        </Dropdown>
+
+                    </p>
+                    <span style={{
+                        fontSize: '12px',
+                        color: '#111',
+                    }}
+                    >
+                        {intl.formatMessage({ id: 'datasource_updateTime' })}
+                        ：
+                        {item.updateTime}
+
+                    </span>
+
+                </div>
+                {/* <Card
                     title={(
-                        <a onClick={() => {
-                            goDetail(item);
-                        }}
+                        <a
+                            style={{
+                                color: '#ffd56a',
+                            }}
+                            onClick={() => {
+                                goDetail(item);
+                            }}
                         >
                             {item.name}
                         </a>
@@ -67,7 +179,13 @@ const Index: React.FC<IndexProps> = ({
                                 )}
                                 placement="bottomLeft"
                             >
-                                <a>{intl.formatMessage({ id: 'common_more' })}</a>
+                                <a style={{
+                                    color: '#ffd56a',
+                                }}
+                                >
+                                    {intl.formatMessage({ id: 'common_more' })}
+
+                                </a>
                             </Dropdown>
                         </div>
                     )}
@@ -85,13 +203,18 @@ const Index: React.FC<IndexProps> = ({
                         {' '}
                         {item.updateTime}
                     </p>
-                </Card>
+                </Card> */}
             </Col>
         </React.Fragment>
     );
     return (
         <>
-            <div className="dv-card-warp">
+            <div
+                className="dv-card-warp"
+                style={{
+                    height: 'calc(100% - 70px)',
+                }}
+            >
                 <Row gutter={16}>
                     {
                         (tableData.list || []).map((item) => (renderItem(item)))
@@ -107,6 +230,25 @@ const Index: React.FC<IndexProps> = ({
                     total={tableData.total}
                 />
             </div>
+            <Modal
+                width="700px"
+                footer={[]}
+                title={intl.formatMessage({ id: 'home_Metadata_crawl_schedule' })}
+                onCancel={() => {
+                    setisScheduleOpen(false);
+                }}
+                open={isScheduleOpen}
+            >
+                <Schedule
+                    onSavaEnd={() => {
+                        setisScheduleOpen(false);
+                    }}
+                    width="100%"
+                    style={{ height: 'auto' }}
+                    jobId={uuid}
+                    isShowPush
+                />
+            </Modal>
         </>
     );
 };
