@@ -2,12 +2,14 @@ import React, { FC } from 'react';
 import './pageContainer.less';
 
 type TPageContainerProps = {
-    footer?: React.ReactNode
+    footer?: React.ReactNode,
+    children: React.ReactNode,
+    style?:any
 }
 
-const PageContainer: FC<TPageContainerProps> = ({ footer, children }) => (
+const PageContainer: FC<TPageContainerProps> = ({ footer, children, style }) => (
     <>
-        <div className="dv-page-container__children">
+        <div className="dv-page-container__children" style={style}>
             {children}
         </div>
         <div className="dv-page-container__footer">
